+++ src/Products/urban/content/licence/Inspection.py
@@ -47,9 +47,7 @@ schema = Schema((
                 'Ticket',
                 'ProjectMeeting',
                 'PatrimonyCertificate',
-                'CODT_NotaryLetter',
-                'CODT_UrbanCertificateOne'
-                'NotaryLetter',
+                'CODT_UrbanCertificateOne',
                 'UrbanCertificateOne',
             ]
         ],