+++ src/Products/urban/content/licence/Ticket.py
@@ -90,9 +90,7 @@ schema = Schema((
                 'Inspection',
                 'ProjectMeeting',
                 'PatrimonyCertificate',
-                'CODT_NotaryLetter',
-                'CODT_UrbanCertificateOne'
-                'NotaryLetter',
+                'CODT_UrbanCertificateOne',
                 'UrbanCertificateOne',
             ]
         ],